packages/expectlib/expectlib-0.2.0.tar.gz/expectlib-0.2.0/expectlib/comparison.py
from typing import Any


class Comparison(object):
    def __init__(
        self,
        value: Any,
        target: Any = None,
        negate: bool = False,
        negate_term: str = ""
    ):
        """
        Test value against target
        :param value: Value being tested
        :param target: Value being tested against
        :param fail_message: Message to display on failure
        :param negate: Negate comparison
        :param negate_term: String in message to represent negation
        """
        self.value = value
        self.target = target
        self.negate: bool = negate
        self.negate_term = negate_term
        self.message_parts = (
            repr(self.value),
            self.negate_term,
            self.name,
            repr(self.target)
        )
        self.on_init(self)

    @staticmethod
    def on_init(comparison):
        """
        Callback function called on initialization
        :return:
        """
        pass

    @staticmethod
    def on_compare(result: bool, message: str):
        """
        Callback function called after comparision is completed
        :param result: Boolean of comparison result
        :param message: String of comparison result
        :return:
        """
        pass

    def compare(self):
        """
        Function called for comparison
        :return: Boolean if comparison passed or failed
        """
        raise NotImplemented()

    @property
    def name(self):
        """
        Normalized name of comparison for message
        :return: Name of comparison
        """
        from expectlib.utils import from_pascal_to_snake

        return from_pascal_to_snake(self.__class__.__name__) \
            .replace("_", " ")

    def __call__(self, fail_message=""):
        """
        Runtime function for comparison
        :return:
        """
        result = self.compare()
        self.called = True

        if self.negate is True:
            result = not result

        message_from_parts = ' '.join(filter(None, self.message_parts))
        message = ": ".join(filter(None, [message_from_parts, fail_message]))

        self.on_compare(result, message)

        assert result, message


class Not(Comparison):
    """
    Mixin to create negated comparisons
    """
    def __init__(self, *args, **kwargs):
        super().__init__(*args, **kwargs)
        self.negate = True


class Equal(Comparison):
    def compare(self):
        if isinstance(self.target, bool) or self.target is None:
            return self.value is self.target
        return self.value == self.target


class DoesntEqual(Equal, Not):
    pass


class GreaterThan(Comparison):
    def compare(self):
        return self.value > self.target


class IsntGreaterThan(GreaterThan, Not):
    pass


class GreaterThanOrEqual(Comparison):
    def compare(self):
        return self.value >= self.target


class IsntGreaterThanOrEqual(GreaterThanOrEqual, Not):
    pass


class LesserThan(Comparison):
    def compare(self):
        return self.value < self.target


class IsntLesserThan(LesserThan, Not):
    pass


class LesserThanOrEqual(Comparison):
    def compare(self):
        result = self.value <= self.target
        return result


class IsntLesserThanOrEqual(LesserThanOrEqual, Not):
    pass


class Empty(Comparison):
    def __init__(self, *args, **kwargs):
        super().__init__(*args, **kwargs)
        self.message_parts = (
            str(self.value),
            self.negate_term,
            self.name
        )

    def compare(self):
        if self.value is None:
            return True

        return len(self.value) == 0


class IsntEmpty(Empty, Not):
    pass


class Contain(Comparison):
    def compare(self):
        return self.target in self.value


class DoesNotContain(Contain, Not):
    pass


class Match(Comparison):
    def compare(self):
        import re
        return re.match(self.target, self.value)


class DoesntMatch(Match, Not):
    pass


class StartWith(Comparison):
    def compare(self):
        text: str = self.value
        return text.startswith(self.target)


class DoesntStartWith(StartWith, Not):
    pass


class EndWith(Comparison):
    def compare(self):
        text: str = self.value
        return text.endswith(self.target)


class DoesntEndWith(EndWith, Not):
    pass


class Raises(Comparison):
    def compare(self):
        try:
            self.value()
        except self.target:
            return True
        else:
            return False


class Callable(Comparison):
    def compare(self):
        return callable(self.value)

class IsntCallable(Callable, Not):
    pass
